Python tars_control_lib_py : better handling of unavailable services.

+++ tars_robot/tars_control_lib_py/src/tars_control_lib_py/gazebo_helper.py
@@ -17,6 +17,10 @@ def gazebo_delete_all_objects():
 
 	print "  Setting up services...",
 	get_world_properties = setupService('/gazebo/get_world_properties', GetWorldProperties)
+
+	if get_world_properties is None:
+		return False
+
 	try:
 		world_properties = get_world_properties()
 		for name in world_properties.model_names:
@@ -29,6 +33,10 @@ def gazebo_delete_all_objects():
 
 def gazebo_delete_object(name):
 	delete_model_service = setupService('/gazebo/delete_model', DeleteModel)
+
+	if delete_model_service is None:
+		return False
+
 	try:
 		print "   Deleting %s ..." % name,
 		result = delete_model_service(DeleteModelRequest(name))	
@@ -53,6 +61,9 @@ def gazebo_spawn_robot(robot_file, name, controllers):
 
 	spawn_model = setupService('/gazebo/spawn_urdf_model', SpawnModel)
 
+	if spawn_model is None:
+		return False
+
 	print "Spawning model..."
 	try:
 		spawn_model(name, model_urdf, name, initial_pose, "world")
@@ -64,6 +75,9 @@ def gazebo_spawn_robot(robot_file, name, controllers):
 
 	load_controller = setupService('/'+name+'/controller_manager/load_controller', LoadController)
 	switch_controllers = setupService('/'+name+'/controller_manager/switch_controller', SwitchController)
+
+	if load_controller is None or switch_controllers is None:
+		return False
 
 	try:
 		for controller in controllers:
@@ -92,22 +106,43 @@ def gazebo_spawn_object(name, sdf_file, x=0, y=0, z=0):
 	model_sdf_f = open(sdf_file, 'r')
 	model_sdf = model_sdf_f.read()
 
-	print "Adding %s @ (%.2f, %.2f, %.2f)" % (name, x, y, z)
+	print "Adding %s @ (%.2f, %.2f, %.2f)" % (name, x, y, z),
 	spawn_sdf_model = setupService('/gazebo/spawn_sdf_model', SpawnModel)
+
+	if spawn_sdf_model is None:
+		print "error : spawn service unavailable."
+		return None
 
 	try:
 		spawn_sdf_model(name, model_sdf, name, initial_pose, "world")
+		print "ok"
 	except Exception as err:
-		print "Error loading cone: %s" % err
+		print "error : %s" % err
 
 
 def gazebo_get_model_state(name, context='world'):
 	get_model_state = setupService('/gazebo/get_model_state', GetModelState)
-	model_state = get_model_state(GetModelStateRequest(name, context))
-	if not model_state.success:
-		return False
-	return model_state 
+
+	if get_model_state is None:
+		return None
+
+	try:
+		model_state = get_model_state(GetModelStateRequest(name, context))
+		if not model_state.success:
+			return False
+		return model_state
+	except Exception as err:
+		print "Error get_model_state: %s" % (name)
+		print err
 
 def gazebo_set_model_state(name, pose, twist, context='world'):
 	set_model_state = setupService('/gazebo/set_model_state', SetModelState)
-	return set_model_state(SetModelStateRequest(ModelState(name, pose, twist, context)))
+
+	if set_model_state is None:
+		return None
+
+	try:
+		return set_model_state(SetModelStateRequest(ModelState(name, pose, twist, context)))
+	except Exception as err:
+		print "Error get_model_state: %s" % (name)
+		print err
